% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{q_catch}
\alias{q_catch}
\title{Query catch data}
\usage{
q_catch(year, db)
}
\arguments{
\item{year}{year(s) of interest}

\item{db}{database connection}
}
\value{
data.frame of catches and vessel info - unfiltered
}
\description{
Query catch data
}
\examples{
\dontrun{
catch_data <- q_catch(2022, db)
}
}
