% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{q_lengths}
\alias{q_lengths}
\title{Query length data}
\usage{
q_lengths(year, db)
}
\arguments{
\item{year}{year(s) of interest}

\item{db}{database connection}
}
\value{
data.frame of lengths and species catch codes
}
\description{
Query length data
}
\examples{
\dontrun{
length_data <- q_length(2022, db)
}
}
