% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{q_strata}
\alias{q_strata}
\title{Query strata data}
\usage{
q_strata(db)
}
\arguments{
\item{db}{database connection}
}
\value{
data.frame of area_stratum, areas, and depth_stratum
}
\description{
Query strata data
}
\examples{
\dontrun{
strata_data <- q_strata(db)
}
}
